feat(gdb): set affinity in threads, fetch vcpu id from object

+++ src/vm.rs
@@ -79,7 +79,7 @@ pub(crate) mod internal {
 		/// Create a new CPU object
 		fn new_cpu(
 			&self,
-			id: u32,
+			id: usize,
 			kernel_info: Arc<KernelInfo>,
 			enable_stats: bool,
 		) -> HypervisorResult<Self::VCPU>;
@@ -443,7 +443,7 @@ impl<VirtBackend: VirtualizationBackend> UhyveVm<VirtBackend> {
 			"gdb is only supported on linux (yet)"
 		);
 
-		let vcpus: Vec<_> = (0..cpu_count)
+		let vcpus: Vec<_> = (0..cpu_count as usize)
 			.map(|cpu_id| {
 				virt_backend
 					.new_cpu(cpu_id, kernel_info.clone(), kernel_info.params.stats)
